Refactor Conjur provider into multiple files
