+++ e2e/suites/provider/cases/conjur/conjur.go
@@ -32,7 +32,7 @@ var _ = Describe("[conjur]", Label("conjur"), func() {
 
 	DescribeTable("sync secrets",
 		framework.TableFuncWithExternalSecret(f, prov),
-		// uses api key auth
+		// use api key auth
 		framework.Compose(withTokenAuth, f, common.FindByName, useApiKeyAuth),
 		framework.Compose(withTokenAuth, f, common.FindByNameAndRewrite, useApiKeyAuth),
 		framework.Compose(withTokenAuth, f, common.FindByTag, useApiKeyAuth),
@@ -53,7 +53,7 @@ var _ = Describe("[conjur]", Label("conjur"), func() {
 		framework.Compose(withTokenAuth, f, common.NestedJSONWithGJSON, useApiKeyAuth),
 		framework.Compose(withTokenAuth, f, common.SyncV1Alpha1, useApiKeyAuth),
 
-		// // use jwt k8s provider
+		// use jwt k8s provider
 		framework.Compose(withJWTK8s, f, common.FindByName, useJWTK8sProvider),
 		framework.Compose(withJWTK8s, f, common.FindByNameAndRewrite, useJWTK8sProvider),
 		framework.Compose(withJWTK8s, f, common.FindByTag, useJWTK8sProvider),
@@ -62,7 +62,7 @@ var _ = Describe("[conjur]", Label("conjur"), func() {
 		framework.Compose(withJWTK8s, f, common.JSONDataFromSync, useJWTK8sProvider),
 		framework.Compose(withJWTK8s, f, common.JSONDataFromRewrite, useJWTK8sProvider),
 
-		// // use jwt k8s hostid provider
+		// use jwt k8s hostid provider
 		framework.Compose(withJWTK8sHostID, f, common.FindByName, useJWTK8sHostIDProvider),
 		framework.Compose(withJWTK8sHostID, f, common.FindByNameAndRewrite, useJWTK8sHostIDProvider),
 		framework.Compose(withJWTK8sHostID, f, common.FindByTag, useJWTK8sHostIDProvider),
