+++ pkg/provider/conjur/provider_test.go
@@ -126,9 +126,9 @@ func TestValidateStore(t *testing.T) {
 			err:   fmt.Errorf("missing Auth.* configuration"),
 		},
 	}
-	c := Provider{}
+	p := Provider{}
 	for _, tc := range testCases {
-		_, err := c.ValidateStore(tc.store)
+		_, err := p.ValidateStore(tc.store)
 		if tc.err != nil && err != nil && err.Error() != tc.err.Error() {
 			t.Errorf("test failed! want %v, got %v", tc.err, err)
 		} else if tc.err == nil && err != nil {
